blog/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
cffi==1.15.1
colorama==0.4.6
comm==0.1.2
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
filelock==3.9.0
ipykernel==6.20.2
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
Markdown==3.4.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2022.7
virtualenv==20.19.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
